% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human1Parser.R
\name{Human1Parser}
\alias{Human1Parser}
\title{Human1Parser}
\value{
the document object, an instance of XMLInternalDocument, XMLAbstractDocument

a names list of counts

long name

two data.frames: species and speciesRefs

a list of data.frames: reaction, reactionrefs, reactants, products, a list of genes
}
\description{
an R6 class which parses xml Human1 reactions into data.frames
}
\details{
import R6
import XML
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Human1Parser$new()}}
\item \href{#method-getDoc}{\code{Human1Parser$getDoc()}}
\item \href{#method-getCounts}{\code{Human1Parser$getCounts()}}
\item \href{#method-getCompartment}{\code{Human1Parser$getCompartment()}}
\item \href{#method-getSpecies}{\code{Human1Parser$getSpecies()}}
\item \href{#method-getReaction}{\code{Human1Parser$getReaction()}}
\item \href{#method-clone}{\code{Human1Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new parser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$new(xml.filename)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a new `Human1Parser` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDoc"></a>}}
\if{latex}{\out{\hypertarget{method-getDoc}{}}}
\subsection{Method \code{getDoc()}}{
retrieve the xml document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getDoc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCounts"></a>}}
\if{latex}{\out{\hypertarget{method-getCounts}{}}}
\subsection{Method \code{getCounts()}}{
learn the elements and their count
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getCounts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCompartment"></a>}}
\if{latex}{\out{\hypertarget{method-getCompartment}{}}}
\subsection{Method \code{getCompartment()}}{
retrieve the long descriptive name of the cellular compartment in which a species is found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getCompartment(shortName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shortName}}{character, e.g. "s"   "p"   "m"   "c"   "l"   "r"   "g"   "n"   "c_i"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSpecies"></a>}}
\if{latex}{\out{\hypertarget{method-getSpecies}{}}}
\subsection{Method \code{getSpecies()}}{
extract one molecular species
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getSpecies(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer the index of the reaction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReaction"></a>}}
\if{latex}{\out{\hypertarget{method-getReaction}{}}}
\subsection{Method \code{getReaction()}}{
extract one reaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getReaction(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer the index of the reaction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
