% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human1Parser.R
\name{Human1Parser}
\alias{Human1Parser}
\title{Human1Parser}
\value{
a names list of counts
}
\description{
an R6 class which parses xml Human1 reactions into data.frames
}
\details{
import R6
import XML
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Human1Parser$new()}}
\item \href{#method-getCounts}{\code{Human1Parser$getCounts()}}
\item \href{#method-clone}{\code{Human1Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new parser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$new(xml.filename)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a new `Human1Parser` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCounts"></a>}}
\if{latex}{\out{\hypertarget{method-getCounts}{}}}
\subsection{Method \code{getCounts()}}{
learn the elements and their count
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$getCounts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human1Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
